BUGFIX: Fix case where settings are empty
The first time that the plugin runs, initial settings are empty.
Settings later get array_merged with new saves, but if the settings
we initially empty, php returns null (thanks PHP); this fix makes
sure that if the initial save isn't there, the defaults are in effect.

Good catch, @Diziara. You get a php cookie!